Add VS code client server & fix launcher package name
