+++ org.eclipse.winery.lsp/META-INF/MANIFEST.MF
@@ -1,0 +1,3 @@
+Manifest-Version: 1.0
+Main-Class: org.eclipse.winery.lsp.Lancher.StdioLauncher
+
