+++ org.eclipse.winery.lsp/src/main/java/org/eclipse/winery/lsp/Client/VSCodeExtention/toscalsp/.gitignore
@@ -1,0 +1,5 @@
+out
+dist
+node_modules
+.vscode-test/
+*.vsix
