+++ org.eclipse.winery.lsp/src/main/java/org/eclipse/winery/lsp/Client/VSCodeExtention/toscalsp/.vscode-test.mjs
@@ -1,0 +1,5 @@
+import { defineConfig } from '@vscode/test-cli';
+
+export default defineConfig({
+	files: 'out/test/**/*.test.js',
+});
