+++ org.eclipse.winery.lsp/src/main/java/org/eclipse/winery/lsp/Client/VSCodeExtention/toscalsp/CHANGELOG.md
@@ -1,0 +1,9 @@
+# Change Log
+
+All notable changes to the "toscalsp" extension will be documented in this file.
+
+Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
+
+## [Unreleased]
+
+- Initial release
