+++ org.eclipse.winery.lsp/src/main/java/org/eclipse/winery/lsp/Client/VSCodeExtention/toscalsp/package.json
@@ -1,0 +1,41 @@
+{
+  "name": "tosca-lsp-client",
+  "displayName": "TOSCA LSP Client",
+  "description": "VS Code client for TOSCA Language Server",
+  "version": "0.0.1",
+  "engines": {
+      "vscode": "^1.50.0"
+  },
+  "categories": [
+      "Other"
+  ],
+  "activationEvents": [
+      "onLanguage:yaml"
+  ],
+  "main": "./out/extension.js",
+  "contributes": {
+      "languages": [
+          {
+              "id": "yaml",
+              "aliases": ["YAML", "yaml"],
+              "extensions": [".yaml", ".yml"]
+          }
+      ]
+  },
+  "scripts": {
+      "vscode:prepublish": "npm run compile",
+      "compile": "tsc -p ./",
+      "watch": "tsc -watch -p ./",
+      "postinstall": "node ./node_modules/vscode/bin/install",
+      "test": "npm run compile && node ./out/test/runTest.js"
+  },
+  "devDependencies": {
+      "@types/node": "^12.11.7",
+      "@types/vscode": "^1.50.0",
+      "typescript": "^4.0.3",
+      "vscode-test": "^1.3.0"
+  },
+  "dependencies": {
+      "vscode-languageclient": "^7.0.0"
+  }
+}
