+++ org.eclipse.winery.lsp/src/main/java/org/eclipse/winery/lsp/Launcher/StdioLauncher.java
@@ -1,0 +1,28 @@
+package org.eclipse.winery.lsp.Launcher;
+import org.eclipse.lsp4j.jsonrpc.Launcher;
+import org.eclipse.lsp4j.launch.LSPLauncher;
+import org.eclipse.lsp4j.services.LanguageClient;
+import org.eclipse.winery.lsp.Server.ServerCore.ToscaLangaugeServer;
+import java.io.InputStream;
+import java.io.OutputStream;
+import java.util.concurrent.ExecutionException;
+import java.util.concurrent.Future;
+import java.util.logging.LogManager;
+import java.util.logging.Logger;
+
+public class StdioLauncher {
+    public static void startServer(InputStream in, OutputStream out)
+        throws InterruptedException, ExecutionException {
+        ToscaLangaugeServer server = new ToscaLangaugeServer();
+        Launcher<LanguageClient> launcher = LSPLauncher.createServerLauncher(server, in, out);
+        server.connect(launcher.getRemoteProxy());
+        Future<?> startListening = launcher.startListening();
+        startListening.get();
+    }
+    public static void main(String[] args) throws ExecutionException, InterruptedException {
+        LogManager.getLogManager().reset();
+        Logger globalLogger = Logger.getLogger(java.util.logging.Logger.GLOBAL_LOGGER_NAME);
+        globalLogger.setLevel(java.util.logging.Level.OFF);
+        startServer(System.in, System.out);
+    }
+}
