+++ org.eclipse.winery.lsp/src/main/java/org/eclipse/winery/lsp/Server/ServerAPI/API/context/BaseOperationContext.java
@@ -1,0 +1,10 @@
+package org.eclipse.winery.lsp.Server.ServerAPI.API.context;
+
+import org.eclipse.lsp4j.ClientCapabilities;
+import org.eclipse.lsp4j.services.LanguageClient;
+import org.eclipse.winery.lsp.Server.ServerAPI.API.ClientLogManager;
+public interface BaseOperationContext {
+    ClientCapabilities clientCapabilities();
+    ClientLogManager clientLogManager();
+    LanguageClient getClient();
+}
