+++ org.eclipse.winery.lsp/src/main/java/org/eclipse/winery/lsp/Server/ServerAPI/API/context/ContextBuilder.java
@@ -1,0 +1,8 @@
+package org.eclipse.winery.lsp.Server.ServerAPI.API.context;
+
+public class ContextBuilder {
+    public static BaseOperationContext baseContext(LSContext serverContext) {
+        return new BaseOperationContextImpl(serverContext);
+    }
+
+}
