--- org.eclipse.winery.lsp/src/main/java/org/eclpse/winery/lsp/Main.java
@@ -1,7 +1,0 @@
-package org.eclpse.winery.lsp;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world AMOR!");
-    }
-}
