--- org.eclipse.winery.lsp/src/main/java/org/eclpse/winery/lsp/Server/ServerAPI/API/context/ContextBuilder.java
@@ -1,8 +1,0 @@
-package org.eclpse.winery.lsp.Server.ServerAPI.API.context;
-
-public class ContextBuilder {
-    public static BaseOperationContext baseContext(LSContext serverContext) {
-        return new BaseOperationContextImpl(serverContext);
-    }
-
-}
