--- org.eclipse.winery.lsp/src/main/java/org/eclpse/winery/lsp/Server/ServerCore/ArtifactTypesDiagnosticsAnalyzer.java
@@ -1,22 +1,0 @@
-package org.eclpse.winery.lsp.Server.ServerCore;
-
-import org.eclipse.lsp4j.Diagnostic;
-import org.eclipse.lsp4j.DiagnosticSeverity;
-import org.eclipse.lsp4j.Position;
-import org.eclipse.lsp4j.PublishDiagnosticsParams;
-import org.eclipse.lsp4j.Range;
-import org.eclipse.lsp4j.services.LanguageClient;
-import org.yaml.snakeyaml.Yaml;
-
-import java.io.InputStream;
-import java.nio.file.Files;
-import java.nio.file.Paths;
-import java.util.ArrayList;
-import java.util.HashMap;
-import java.util.List;
-import java.util.Map;
-import java.util.concurrent.ConcurrentHashMap;
-public class ArtifactTypesDiagnosticsAnalyzer {
-    
-    
-}
