+++ org.eclipse.winery.lsp/src/test/resources/testArtifactType.yaml
@@ -1,0 +1,5 @@
+    description: >
+      Artifact for the interpreted Python language.
+    derived_from: Implementation
+    mime_typee: application/x-python
+    file_ext: [ py ]
